refactor(runtime): split methods_string.rs into cohesive modules (§7-8)
Split the 2014-line methods_string.rs (over the 500-line limit) into
seven cohesive modules, all under 500 lines:

- methods_string.rs (462)        - subst core: SubstCaseTransforms,
                                   nth helpers, dispatch_subst
- methods_string_encoding.rs (376) - encode/decode entry points
- methods_string_codec.rs (143)  - encoding_rs codec helpers
- methods_string_subst_repl.rs (133) - eval_subst_replacement_cased
- methods_string_search.rs (171) - contains/starts-with/ends-with
- methods_string_index.rs (398)  - index/indices/rindex + helpers
- methods_string_substr.rs (357) - substr/substr-rw/substr-eq

All methods are impl Interpreter, so the split is visibility-only: a
few cross-module private fns (apply, char_is_encodable,
decode_utf8_with_replacement, runtime_error_to_failure) are bumped to
pub(super), and the subst-repl module imports SubstCaseTransforms.
No behavior change.
