+++ src/runtime/mod.rs
@@ -209,6 +209,12 @@ mod methods_signature_candidates;
 pub(crate) mod methods_signature_errors;
 mod methods_signature_shaped;
 mod methods_string;
+mod methods_string_codec;
+mod methods_string_encoding;
+mod methods_string_index;
+mod methods_string_search;
+mod methods_string_subst_repl;
+mod methods_string_substr;
 mod methods_sub;
 mod methods_supply_dispatch;
 mod methods_temporal;
